test(scenario-designer): validation refacto
